feat: centralize shared restic utilities
